Add NativeStorage class
