+++ osu.Framework.Tests/IO/TestDesktopStorage.cs
@@ -15,7 +15,7 @@ namespace osu.Framework.Tests.IO
         public void TestRelativePaths()
         {
             var guid = new Guid().ToString();
-            var storage = new DesktopStorage(guid);
+            var storage = new NativeStorage(guid);
 
             var basePath = storage.GetFullPath(string.Empty);
