+++ osu.Framework.iOS/IOSGameHost.cs
@@ -10,7 +10,6 @@ using osu.Framework.IO.Stores;
 using osu.Framework.iOS.Graphics.Textures;
 using osu.Framework.iOS.Input;
 using osu.Framework.Platform;
-using osu.Framework.Platform.MacOS;
 
 namespace osu.Framework.iOS
 {
@@ -39,7 +38,7 @@ namespace osu.Framework.iOS
         protected override IEnumerable<InputHandler> CreateAvailableInputHandlers() =>
             new InputHandler[] { new IOSTouchHandler(gameView), new IOSKeyboardHandler(gameView), new IOSRawKeyboardHandler() };
 
-        protected override Storage GetStorage(string baseName) => new MacOSStorage(baseName, this);
+        protected override Storage GetStorage(string baseName) => new IOSStorage(baseName, this);
 
         public override void OpenFileExternally(string filename) => throw new NotImplementedException();
