+++ osu.Framework/Platform/MacOS/MacOSStorage.cs
@@ -8,7 +8,7 @@ namespace osu.Framework.Platform.MacOS
 {
     public class MacOSStorage : DesktopStorage
     {
-        public MacOSStorage(string baseName, GameHost host)
+        public MacOSStorage(string baseName, DesktopGameHost host)
             : base(baseName, host)
         {
         }
